[reconfigurator] Move planner tests to their corresponding submodules
