+++ nexus/reconfigurator/planning/src/mgs_updates/rot.rs
@@ -289,3 +289,313 @@ pub fn try_make_update_rot(
         artifact_version: artifact.id.version.clone(),
     })
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::mgs_updates::ImpossibleUpdatePolicy;
+    use crate::mgs_updates::PlannedMgsUpdates;
+    use crate::mgs_updates::plan_mgs_updates;
+    use crate::mgs_updates::test_helpers::ARTIFACT_HASH_ROT_GIMLET_B;
+    use crate::mgs_updates::test_helpers::ARTIFACT_HASH_ROT_SWITCH_B;
+    use crate::mgs_updates::test_helpers::ARTIFACT_VERSION_1;
+    use crate::mgs_updates::test_helpers::ARTIFACT_VERSION_1_5;
+    use crate::mgs_updates::test_helpers::ARTIFACT_VERSION_2;
+    use crate::mgs_updates::test_helpers::TestBoards;
+    use dropshot::ConfigLogging;
+    use dropshot::ConfigLoggingLevel;
+    use dropshot::test_util::LogContext;
+    use gateway_client::types::SpType;
+    use gateway_types::rot::RotSlot;
+    use nexus_types::deployment::ExpectedVersion;
+    use nexus_types::deployment::PendingMgsUpdateDetails;
+    use nexus_types::deployment::PendingMgsUpdateRotDetails;
+    use nexus_types::deployment::PendingMgsUpdates;
+    use nexus_types::deployment::TargetReleaseDescription;
+    use std::collections::BTreeSet;
+
+    // Short hand-rolled update sequence that exercises some basic behavior for
+    // RoT updates.
+    #[test]
+    fn test_basic_rot() {
+        let test_name = "planning_mgs_updates_basic_rot";
+        let logctx = LogContext::new(
+            test_name,
+            &ConfigLogging::StderrTerminal { level: ConfigLoggingLevel::Debug },
+        );
+        let log = &logctx.log;
+        let test_boards = TestBoards::new(test_name);
+
+        // Test that with no updates pending and no TUF repo specified, there
+        // will remain no updates pending.
+        let collection = test_boards
+            .collection_builder()
+            .rot_active_version_exception(SpType::Sled, 0, ARTIFACT_VERSION_1)
+            .build();
+        let current_boards = &collection.baseboards;
+        let initial_updates = PendingMgsUpdates::new();
+        let nmax_updates = 1;
+        let impossible_update_policy = ImpossibleUpdatePolicy::Reevaluate;
+        let PlannedMgsUpdates { pending_updates: updates, .. } =
+            plan_mgs_updates(
+                log,
+                &collection,
+                current_boards,
+                &initial_updates,
+                &TargetReleaseDescription::Initial,
+                nmax_updates,
+                impossible_update_policy,
+            );
+        assert!(updates.is_empty());
+
+        // Test that when a TUF repo is specified and one RoT is outdated, then
+        // it's configured with an update (and the update looks correct).
+        let repo = test_boards.tuf_repo();
+        let PlannedMgsUpdates { pending_updates: updates, .. } =
+            plan_mgs_updates(
+                log,
+                &collection,
+                current_boards,
+                &initial_updates,
+                &TargetReleaseDescription::TufRepo(repo.clone()),
+                nmax_updates,
+                impossible_update_policy,
+            );
+        assert_eq!(updates.len(), 1);
+        let first_update = updates.iter().next().expect("at least one update");
+        assert_eq!(first_update.baseboard_id.serial_number, "sled_0");
+        assert_eq!(first_update.sp_type, SpType::Sled);
+        assert_eq!(first_update.slot_id, 0);
+        assert_eq!(first_update.artifact_hash, ARTIFACT_HASH_ROT_GIMLET_B);
+        assert_eq!(first_update.artifact_version, ARTIFACT_VERSION_2);
+
+        // Test that when an update is already pending, and nothing changes
+        // about the state of the world (i.e., the inventory), then the planner
+        // makes no changes.
+        let PlannedMgsUpdates { pending_updates: later_updates, .. } =
+            plan_mgs_updates(
+                log,
+                &collection,
+                current_boards,
+                &updates,
+                &TargetReleaseDescription::TufRepo(repo.clone()),
+                nmax_updates,
+                impossible_update_policy,
+            );
+        assert_eq!(updates, later_updates);
+
+        // Test that when two updates are needed, but one is already pending,
+        // then the other one is *not* started (because it exceeds
+        // nmax_updates).
+        let later_collection = test_boards
+            .collection_builder()
+            .rot_active_version_exception(SpType::Sled, 0, ARTIFACT_VERSION_1)
+            .rot_active_version_exception(SpType::Switch, 1, ARTIFACT_VERSION_1)
+            .build();
+        let PlannedMgsUpdates { pending_updates: later_updates, .. } =
+            plan_mgs_updates(
+                log,
+                &later_collection,
+                current_boards,
+                &updates,
+                &TargetReleaseDescription::TufRepo(repo.clone()),
+                nmax_updates,
+                impossible_update_policy,
+            );
+        assert_eq!(updates, later_updates);
+
+        // At this point, we're ready to test that when the first SpType update
+        // completes, then the second one *is* started.  This tests four
+        // different things: first that we noticed the first one completed,
+        // second that we noticed another thing needed an update, third that
+        // an update is correctly configured after the active slot has changed
+        // from the previous component update, and fourth that the planner
+        // schedules the updates in the correct order: first RoT, and second SP.
+        let later_collection = test_boards
+            .collection_builder()
+            .sp_active_version_exception(SpType::Switch, 1, ARTIFACT_VERSION_1)
+            .rot_active_version_exception(SpType::Switch, 1, ARTIFACT_VERSION_1)
+            .rot_active_slot_exception(SpType::Sled, 0, RotSlot::B)
+            .rot_persistent_boot_preference_exception(
+                SpType::Sled,
+                0,
+                RotSlot::B,
+            )
+            .build();
+        let PlannedMgsUpdates { pending_updates: later_updates, .. } =
+            plan_mgs_updates(
+                log,
+                &later_collection,
+                current_boards,
+                &updates,
+                &TargetReleaseDescription::TufRepo(repo.clone()),
+                nmax_updates,
+                impossible_update_policy,
+            );
+        assert_eq!(later_updates.len(), 1);
+        let next_update =
+            later_updates.iter().next().expect("at least one update");
+        assert_ne!(first_update, next_update);
+        assert_eq!(next_update.baseboard_id.serial_number, "switch_1");
+        assert_eq!(next_update.sp_type, SpType::Switch);
+        assert_eq!(next_update.slot_id, 1);
+        assert_eq!(next_update.artifact_hash, ARTIFACT_HASH_ROT_SWITCH_B);
+        assert_eq!(next_update.artifact_version, ARTIFACT_VERSION_2);
+
+        // Finally, test that when all components are in spec, then no updates
+        // are configured.
+        let updated_collection = test_boards.collection_builder().build();
+        let PlannedMgsUpdates { pending_updates: later_updates, .. } =
+            plan_mgs_updates(
+                log,
+                &updated_collection,
+                current_boards,
+                &later_updates,
+                &TargetReleaseDescription::TufRepo(repo.clone()),
+                nmax_updates,
+                impossible_update_policy,
+            );
+        assert!(later_updates.is_empty());
+
+        // Test that we don't try to update boards that aren't in
+        // `current_boards`, even if they're in inventory and outdated.
+        let collection = test_boards
+            .collection_builder()
+            .rot_active_version_exception(SpType::Sled, 0, ARTIFACT_VERSION_1)
+            .build();
+        let PlannedMgsUpdates { pending_updates: updates, .. } =
+            plan_mgs_updates(
+                log,
+                &collection,
+                &BTreeSet::new(),
+                &PendingMgsUpdates::new(),
+                &TargetReleaseDescription::TufRepo(repo.clone()),
+                nmax_updates,
+                impossible_update_policy,
+            );
+        assert!(updates.is_empty());
+        let PlannedMgsUpdates { pending_updates: updates, .. } =
+            plan_mgs_updates(
+                log,
+                &collection,
+                &collection.baseboards,
+                &PendingMgsUpdates::new(),
+                &TargetReleaseDescription::TufRepo(repo.clone()),
+                nmax_updates,
+                impossible_update_policy,
+            );
+        // We verified most of the details above.  Here we're just double
+        // checking that the baseboard being missing is the only reason that no
+        // update was generated.
+        assert_eq!(updates.len(), 1);
+
+        // Verify the precondition details of an ordinary RoT update.
+        let old_update =
+            updates.into_iter().next().expect("at least one update");
+        let PendingMgsUpdateDetails::Rot(PendingMgsUpdateRotDetails {
+            expected_active_slot: old_expected_active_slot,
+            expected_inactive_version: old_expected_inactive_version,
+            ..
+        }) = &old_update.details
+        else {
+            panic!("expected RoT update");
+        };
+        assert_eq!(ARTIFACT_VERSION_1, old_expected_active_slot.version());
+        assert_eq!(
+            ExpectedVersion::NoValidVersion,
+            *old_expected_inactive_version
+        );
+
+        // Test that if the inactive slot contents have changed, then we'll get
+        // a new update reflecting that.
+        let collection = test_boards
+            .collection_builder()
+            .sp_versions(
+                ARTIFACT_VERSION_2,
+                ExpectedVersion::Version(ARTIFACT_VERSION_1),
+            )
+            .rot_versions(
+                ARTIFACT_VERSION_2,
+                ExpectedVersion::Version(ARTIFACT_VERSION_1),
+            )
+            .stage0_versions(
+                ARTIFACT_VERSION_2,
+                ExpectedVersion::Version(ARTIFACT_VERSION_1),
+            )
+            .rot_active_version_exception(SpType::Sled, 0, ARTIFACT_VERSION_1)
+            .build();
+        let PlannedMgsUpdates { pending_updates: new_updates, .. } =
+            plan_mgs_updates(
+                log,
+                &collection,
+                &collection.baseboards,
+                &updates,
+                &TargetReleaseDescription::TufRepo(repo.clone()),
+                nmax_updates,
+                impossible_update_policy,
+            );
+        assert_ne!(updates, new_updates);
+        assert_eq!(new_updates.len(), 1);
+        let new_update =
+            new_updates.into_iter().next().expect("at least one update");
+        assert_eq!(old_update.baseboard_id, new_update.baseboard_id);
+        assert_eq!(old_update.sp_type, new_update.sp_type);
+        assert_eq!(old_update.slot_id, new_update.slot_id);
+        assert_eq!(old_update.artifact_hash, new_update.artifact_hash);
+        assert_eq!(old_update.artifact_version, new_update.artifact_version);
+        let PendingMgsUpdateDetails::Rot(PendingMgsUpdateRotDetails {
+            expected_active_slot: new_expected_active_slot,
+            expected_inactive_version: new_expected_inactive_version,
+            ..
+        }) = &new_update.details
+        else {
+            panic!("expected RoT update");
+        };
+        assert_eq!(ARTIFACT_VERSION_1, new_expected_active_slot.version());
+        assert_eq!(
+            ExpectedVersion::Version(ARTIFACT_VERSION_1),
+            *new_expected_inactive_version
+        );
+
+        // Test that if instead it's the active slot whose contents have changed
+        // to something other than the new expected version, then we'll also get
+        // a new update reflecting that.
+        let collection = test_boards
+            .collection_builder()
+            .rot_active_version_exception(SpType::Sled, 0, ARTIFACT_VERSION_1_5)
+            .build();
+        let PlannedMgsUpdates { pending_updates: new_updates, .. } =
+            plan_mgs_updates(
+                log,
+                &collection,
+                &collection.baseboards,
+                &updates,
+                &TargetReleaseDescription::TufRepo(repo.clone()),
+                nmax_updates,
+                impossible_update_policy,
+            );
+        assert_ne!(updates, new_updates);
+        assert_eq!(new_updates.len(), 1);
+        let new_update =
+            new_updates.into_iter().next().expect("at least one update");
+        assert_eq!(old_update.baseboard_id, new_update.baseboard_id);
+        assert_eq!(old_update.sp_type, new_update.sp_type);
+        assert_eq!(old_update.slot_id, new_update.slot_id);
+        assert_eq!(old_update.artifact_hash, new_update.artifact_hash);
+        assert_eq!(old_update.artifact_version, new_update.artifact_version);
+        let PendingMgsUpdateDetails::Rot(PendingMgsUpdateRotDetails {
+            expected_active_slot: new_expected_active_slot,
+            expected_inactive_version: new_expected_inactive_version,
+            ..
+        }) = &new_update.details
+        else {
+            panic!("expected RoT update");
+        };
+        assert_eq!(ARTIFACT_VERSION_1_5, new_expected_active_slot.version());
+        assert_eq!(
+            ExpectedVersion::NoValidVersion,
+            *new_expected_inactive_version
+        );
+
+        logctx.cleanup_successful();
+    }
+}
